only write to db when it is empty
only write to db when it is empty

+++ build.sh
@@ -2,9 +2,11 @@
 
 cd back_end
 npm install
-npx prisma migrate dev
-cd ../data_set
-unzip crypto-dataset.zip
-cd ..
-node back_end/src/utils/writeDataIntoDB.js
-rm -rf data_set/*.csv
+if [ "$DB_EMPTY" = "TRUE" ]; then
+    npx prisma migrate dev
+    cd ../data_set
+    unzip crypto-dataset.zip
+    cd ..
+    node back_end/src/utils/writeDataIntoDB.js
+    rm -rf data_set/*.csv
+fi
